Pigweed-app example: nrfconnect SDK logs redirect to pw_hdlc (#4775)
