+++ examples/pigweed-app/nrfconnect/.gitignore
@@ -1,0 +1,1 @@
+/build/
