+++ examples/pigweed-app/nrfconnect/CMakeLists.txt
@@ -19,6 +19,8 @@ get_filename_component(CHIP_ROOT ${CMAKE_CURRENT_SOURCE_DIR}/third_party/connect
 get_filename_component(NRFCONNECT_COMMON ${CHIP_ROOT}/examples/platform/nrfconnect REALPATH)
 set(PIGWEED_ROOT "${CHIP_ROOT}/third_party/pigweed/repo")
 
+set(CONF_FILE ${CHIP_ROOT}/config/nrfconnect/app/sample-defaults.conf prj.conf)
+
 if(${BOARD} STREQUAL "nrf52840dongle_nrf52840")
     list(INSERT OVERLAY_CONFIG 0 ${CHIP_ROOT}/config/nrfconnect/app/overlay-usb_support.conf)
 endif()
@@ -29,17 +31,47 @@ find_package(Zephyr HINTS $ENV{ZEPHYR_BASE})
 
 project(chip-nrf52840-pigweed-example)
 
-target_include_directories(app PRIVATE main/include 
+zephyr_compile_options(-Werror)
+
+include(${PIGWEED_ROOT}/pw_build/pigweed.cmake)
+include(${PIGWEED_ROOT}/pw_protobuf_compiler/proto.cmake)
+
+pw_set_backend(pw_log pw_log_basic)
+pw_set_backend(pw_assert pw_assert_log)
+pw_set_backend(pw_sys_io pw_sys_io.nrfconnect)
+set(dir_pw_third_party_nanopb PRESENT CACHE STRING ${CHIP_ROOT}/third_party/nanopb/repo FORCE)
+
+add_subdirectory(third_party/connectedhomeip/examples/platform/nrfconnect/pw_sys_io)
+add_subdirectory(third_party/connectedhomeip/third_party/nanopb/repo)
+add_subdirectory(third_party/connectedhomeip/third_party/pigweed/repo)
+
+target_include_directories(app PRIVATE main/include
+                           ${ZEPHYR_BASE}/subsys // required to get access to `log_output_std.h`
                            ${NRFCONNECT_COMMON}/util/include
+                           ${CHIP_ROOT}/src/lib/support
                            ${PIGWEED_ROOT}/pw_sys_io/public
                            ${PIGWEED_ROOT}/pw_span/public_overrides
                            ${PIGWEED_ROOT}/pw_span/public
                            ${PIGWEED_ROOT}/pw_polyfill/public
                            ${PIGWEED_ROOT}/pw_polyfill/standard_library_public
                            ${PIGWEED_ROOT}/pw_polyfill/public_overrides
+                           ${PIGWEED_ROOT}/pw_rpc/public
                            ${PIGWEED_ROOT}/pw_status/public
                            ${PIGWEED_ROOT}/pw_preprocessor/public
+                           ../../common
+                           ../../common/pigweed/nrfconnect
                            ${NRFCONNECT_COMMON}/pw_sys_io/public)
 
-target_sources(app PRIVATE main/main.cpp 
-               ${NRFCONNECT_COMMON}/util/LEDWidget.cpp)
+target_sources(app PRIVATE main/main.cpp
+  ../../common/pigweed/RpcService.cpp
+  ../../common/pigweed/nrfconnect/PigweedLoggerMutex.cpp
+  ${NRFCONNECT_COMMON}/util/LEDWidget.cpp
+  ${NRFCONNECT_COMMON}/util/PigweedLogger.cpp
+)
+
+target_link_libraries(app PUBLIC
+  pw_hdlc
+  pw_log
+  pw_rpc.nanopb.echo_service
+  pw_rpc.server
+)
