+++ examples/pigweed-app/nrfconnect/main/main.cpp
@@ -17,7 +17,11 @@
 
 #include "AppConfig.h"
 #include "LEDWidget.h"
+#include "PigweedLoggerMutex.h"
+#include "pigweed/RpcService.h"
+#include <kernel.h>
 
+#include "pw_rpc/echo_service_nanopb.h"
 #include "pw_sys_io/sys_io.h"
 #include "pw_sys_io_nrfconnect/init.h"
 
@@ -28,9 +32,26 @@ LOG_MODULE_REGISTER(app);
 
 static LEDWidget sStatusLED;
 
-namespace hdlc_example {
-extern void Start();
-} // namespace hdlc_example
+namespace {
+#define RPC_STACK_SIZE (8 * 1024)
+#define RPC_PRIORITY 7
+
+K_THREAD_STACK_DEFINE(rpc_stack_area, RPC_STACK_SIZE);
+struct k_thread rpc_thread_data;
+
+pw::rpc::EchoService echo_service;
+
+void RegisterServices(pw::rpc::Server & server)
+{
+    server.RegisterService(echo_service);
+}
+
+void RunRpcService(void *, void *, void *)
+{
+    Start(RegisterServices, &::chip::rpc::logger_mutex);
+}
+
+} // namespace
 
 int main()
 {
@@ -51,7 +72,8 @@ int main()
     sStatusLED.Init(SYSTEM_STATE_LED);
     sStatusLED.Set(true);
 
-    hdlc_example::Start();
-
+    k_thread_create(&rpc_thread_data, rpc_stack_area, K_THREAD_STACK_SIZEOF(rpc_stack_area), RunRpcService, NULL, NULL, NULL,
+                    RPC_PRIORITY, 0, K_NO_WAIT);
+    k_thread_join(&rpc_thread_data, K_FOREVER);
     return 0;
 }
