+++ examples/pigweed-app/nrfconnect/prj.conf
@@ -13,57 +13,8 @@
 #    See the License for the specific language governing permissions and
 #    limitations under the License.
 #
-CONFIG_CHIP=y
-
-CONFIG_NEWLIB_LIBC=y
-
 # Export POSIX names for functions implementing a subset of POSIX standard in Zephyr
 CONFIG_POSIX_API=y
-CONFIG_PTHREAD_IPC=y
-CONFIG_EVENTFD=y
-
-CONFIG_LOG=n
-CONFIG_LOG_MINIMAL=n
-CONFIG_ASSERT=y
-CONFIG_HW_STACK_PROTECTION=y
-
-# Generic networking options
-CONFIG_NETWORKING=y
-CONFIG_NET_SOCKETS=y
-CONFIG_NET_SOCKETS_POSIX_NAMES=n
-
-# Application stack size
-CONFIG_MAIN_STACK_SIZE=8192
-CONFIG_INIT_STACKS=y
-
-# Disable certain parts of Zephyr IPv6 stack
-CONFIG_NET_IPV6_NBR_CACHE=n
-CONFIG_NET_IPV6_MLD=n
-
-# Use mbedTLS from nrf_security library
-CONFIG_NORDIC_SECURITY_BACKEND=y
-CONFIG_OPENTHREAD_MBEDTLS=n
-
-CONFIG_MBEDTLS_ENABLE_HEAP=y
-CONFIG_MBEDTLS_HEAP_SIZE=15360
-CONFIG_MBEDTLS_TLS_LIBRARY=y
-CONFIG_NRF_SECURITY_ADVANCED=y
-
-CONFIG_MBEDTLS_AES_C=y
-CONFIG_MBEDTLS_RSA_C=y
-CONFIG_MBEDTLS_ECP_C=y
-CONFIG_MBEDTLS_ECP_DP_SECP256R1_ENABLED=y
-CONFIG_MBEDTLS_CTR_DRBG_C=y
-CONFIG_MBEDTLS_CIPHER_MODE_CTR=y
-CONFIG_MBEDTLS_ECJPAKE_C=y
-
-# Disable unneeded crypto operations
-CONFIG_MBEDTLS_SHA512_C=n
-CONFIG_MBEDTLS_CIPHER_MODE_XTS=n
-CONFIG_MBEDTLS_CHACHA20_C=n
-CONFIG_MBEDTLS_POLY1305_C=n
-CONFIG_MBEDTLS_CHACHAPOLY_C=n
-CONFIG_MBEDTLS_GCM_C=n
 
 # Add support for LEDs and buttons on Nordic development kits
 CONFIG_DK_LIBRARY=y
@@ -76,9 +27,23 @@ CONFIG_MPU_STACK_GUARD=y
 CONFIG_CHIP_PW_RPC=y
 
 # Add support for C++17 to build Pigweed components
+CONFIG_STD_CPP14=n
 CONFIG_STD_CPP17=y
 
 # Add support for Zephyr console component to use it for Pigweed console purposes
 CONFIG_CONSOLE_SUBSYS=y
 CONFIG_CONSOLE_GETCHAR=y
-CONFIG_CONSOLE_GETCHAR_BUFSIZE=256
+CONFIG_CONSOLE_PUTCHAR_BUFSIZE=256
+
+# Disable features which may interfere with Pigweed HDLC transport
+CONFIG_SHELL=n
+CONFIG_OPENTHREAD_SHELL=n
+CONFIG_BOOT_BANNER=n
+
+# Configure Zephyr logger with defaults backends disabled as the app provides its own,
+# based on Pigweed HDLC.
+CONFIG_LOG=y
+CONFIG_LOG_MINIMAL=n
+CONFIG_LOG_IMMEDIATE=y
+CONFIG_LOG_BACKEND_UART=n
+CONFIG_LOG_BACKEND_RTT=n
